rename entropy_heatmap to entropy_analysis
